Runtime primitives
